[setup] deletes old server.js, deletes gruntfile.js, moves bcrypt implementation to userController
