+++ server/controllers/userCtrl.js
@@ -5,12 +5,26 @@ var db = require('../db/database.js');
 
 module.exports = {
   signup: function(req, res) {
-    console.log('inside userCtrl.signup');
-    db.users.insert(req.body, function(err, doc) {
-      if(err) {
+    var password = req.body.password;
+    var username = req.body.username;
+
+    var salt = bcrypt.genSaltSync(10);
+    var hash = bcrypt.hashSync(password, salt);
+    req.body.password = hash;
+
+    db.users.find({username: username}, function(err, exists){
+      if(!exists.length){
+        db.users.insert(req.body, function(err, doc){
+          if(err){
+            console.log(err);
+          }
+        });
+      }
+      else {
         console.log(err);
       }
     });
+
     res.send('/signin');
   },
 
@@ -19,22 +33,28 @@ module.exports = {
   },
 
   signin: function(req, res) {
-    var email = req.body.email;
+    var username = req.body.username;
     var password = req.body.password;
 
-    db.users.find({email:email, password: password}, function(err, exists){
+    db.users.find({username: username}, function(err, exists){
       if(!exists.length){
         res.send(false);
       }
 
       else {
-        var payload = { email: email, password: password};
-        var secret = 'brewed';
+        if(bcrypt.compareSync( password, exists[0].password)){
+          var payload = { username: username, password: password};
+          var secret = 'brewed';
 
-        // encode token
-        var token = jwt.encode(payload, secret);
+          // encode token
+          var token = jwt.encode(payload, secret);
 
-        res.send(token);
+          res.send(token);
+        }
+        else{
+          res.send(false);
+        }
+
       }
     });
   }
